+++ pkgs/checks/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix a bug when using asynchronous conditions with `mayEmit` or
   `mayEmitMultiple`. Note that extensions using `nestAsync` should synchronously
   forward exceptions from that call.
+- Output more compact failure messages for simple expectations.
 
 ## 0.3.1